requirements.txt
# Core Agno dependencies
agno>=0.1.0

# FastAPI and server dependencies
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0

# Environment management
python-dotenv>=1.0.0

# Model providers
google-generativeai>=0.8.0

# Tools and integrations
tavily-python>=0.5.0

# Database
sqlalchemy>=2.0.0
